% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js4shiny-package.R
\docType{data}
\name{first_sentences}
\alias{first_sentences}
\title{First Sentences of Books}
\format{
A data frame with 40 rows and 3 variables:
\describe{
\item{\code{quote}}{The first sentence of the book}
\item{\code{title}}{The title of the book}
\item{\code{author}}{The author of the book}
}
}
\usage{
first_sentences
}
\description{
First sentences of books, downloaded from
\url{https://firstsentencesofbooks.tumblr.com/}.
}
\keyword{datasets}
