% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{html_dependency_js4shiny}
\alias{html_dependency_js4shiny}
\alias{html_dependency_redirectConsoleLog}
\alias{html_dependency_stylize}
\title{js4shiny HTML Dependencies}
\usage{
html_dependency_js4shiny(redirectConsole = TRUE, jsonview = TRUE,
  stylize = "all", use_google_fonts = FALSE)

html_dependency_redirectConsoleLog()

html_dependency_stylize(...)
}
\arguments{
\item{redirectConsole}{Include JS and CSS assets to enable literate
programming with JavaScript by redirecting \code{console.log()} to a chunk-
specific output div.}

\item{jsonview}{Include JS and CSS assets to enable tree view display of
JSON objects for the \pkg{knitr} json engine.}

\item{stylize}{One of "none", "all", "fonts", "variables", "table",
"utility", "code", "pandoc-line-numbers" to include the CSS styles
developed for \pkg{js4shiny}.}

\item{use_google_fonts}{Should fonts hosted on Google Fonts be included?
Default is \code{FALSE}, where only system fonts will be used.}

\item{...}{Arguments passed to \code{html_dependency_js4shiny()}.}
}
\description{
Include the various HTML JavaScript and CSS assets created for
\pkg{js4shiny}.
}
\section{Functions}{
\itemize{
\item \code{html_dependency_redirectConsoleLog}: Include just the console redirection
dependencies.

\item \code{html_dependency_stylize}: Include the full or partial
\pkg{js4shiny} CSS styles.
}}

\seealso{
Other js4shiny HTML dependencies: \code{\link{html_setup}}
}
\concept{js4shiny HTML dependencies}
