% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document_plain.R
\name{html_document_plain}
\alias{html_document_plain}
\alias{include_script}
\title{Plain (Minimal) HTML Document}
\usage{
html_document_plain(..., css = "normalize", script = NULL,
  highlight = "haddock", fig_width = 10, fig_height = 7,
  fig_retina = 2, keep_md = FALSE, dev = "png", pandoc_args = NULL,
  extra_dependencies = NULL)

include_script(head = NULL, before = NULL, after = NULL)
}
\arguments{
\item{...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}

\item{css}{A list of css files to include in the document's `<head>`. Include
`"normalize"` in the list of css files to include `normalize.css`, which
provides basic style resetting.}

\item{script}{A list of `.js` files to include in the document. Use
`include_script()` to choose where in the HTML the script should be
sourced.}

\item{highlight}{One of the pandoc highlight styles.}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{fig_retina}{Scaling to perform for retina displays (defaults to 2, which
currently works for all widely used retina displays). Set to \code{NULL} to
prevent retina scaling. Note that this will always be \code{NULL} when
\code{keep_md} is specified (this is because \code{fig_retina} relies on
outputting HTML directly into the markdown document).}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}

\item{head, before, after}{A character vector of source files, each to be
included in the `<head>`, or *before* or *after* the HTML content in
`<body>`.}
}
\description{
This RMarkdown output format provides a minimal HTML5 template and minimal
features for including CSS and JavaScript files in the output source.
}
\section{Functions}{
\itemize{
\item \code{include_script}: Helper function for including JS scripts
}}

