% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{html_setup}
\alias{html_setup}
\alias{html_setup_blogdown}
\title{Load js4shiny HTML assets and knitr dependencies}
\usage{
html_setup(stylize = "none")

html_setup_blogdown(stylize = c("fonts", "variables", "table", "code",
  "utility"))
}
\arguments{
\item{stylize}{One of "none", "all", "fonts", "variables", "table",
"utility", "code", "pandoc-line-numbers" to include the CSS styles
developed for \pkg{js4shiny}.}
}
\description{
Overrides the JavaScript knitr engine, registers knitr output hooks, and
declares the JS and CSS dependencies that are required to enable the
literate JavaScript code chunks inside R Markdown formats that write to HTML.
}
\section{Functions}{
\itemize{
\item \code{html_setup_blogdown}: A blogdown-specific HTML setup that includes styles
for `<pre>` code blocks, tables, some utility functions, and the CSS
variables declaring the \pkg{js4shiny} colors.
}}

\seealso{
Other js4shiny HTML dependencies: \code{\link{html_dependency_js4shiny}}
}
\concept{js4shiny HTML dependencies}
