% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js4shiny-package.R
\docType{package}
\name{js4shiny-package}
\alias{js4shiny}
\alias{js4shiny-package}
\title{js4shiny: Companion Package for JavaScript for Shiny Users}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Companion Package for JavaScript for Shiny Users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/js4shiny}
  \item Report bugs at \url{https://github.com/gadenbuie/js4shiny/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
