% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document.R
\name{js4shiny_rmd}
\alias{js4shiny_rmd}
\title{Create a new js4shiny HTML document}
\usage{
js4shiny_rmd(
  type = c("plain", "js"),
  full_template = FALSE,
  path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{type}{One of \code{"plain"} for \code{\link[=html_document_plain]{html_document_plain()}} or \code{"js"} for
\code{\link[=html_document_js]{html_document_js()}}.}

\item{full_template}{Include the full R Markdown template document. Default
is \code{FALSE}.}

\item{path}{If \code{NULL}, an R Markdown document is opened in a new RStudio
editor. If a path is given, a file is created and opened if in RStudio.}

\item{overwrite}{If \code{TRUE}, will overwrite \code{path} if it exists.}
}
\description{
Opens or creates an R Markdown document using the \pkg{js4shiny} html
document templates.
}
\examples{

tmpfile <- tempfile(fileext = ".Rmd")
js4shiny_rmd(type = "plain", full_template = TRUE, path = tmpfile)
js4shiny_rmd(type = "plain", path = tmpfile, overwrite = TRUE)

}
\seealso{
\code{\link[=html_document_plain]{html_document_plain()}}, \code{\link[=html_document_js]{html_document_js()}}
}
