% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{launch_shiny_in}
\alias{launch_shiny_in}
\title{Choose Launch Location for Shiny Apps}
\usage{
launch_shiny_in(where = NULL)
}
\arguments{
\item{where}{One of \code{"external"}, \code{"pane"}, or \code{"window"}.}
}
\description{
This function sets the \code{shiny.launch.browser} option to launch Shiny apps in
an \code{"external"} browser, the RStudio viewer \code{"pane"}, or a new \code{"window"} in
RStudio.
}
