% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{lint_js_addin}
\alias{lint_js_addin}
\title{Lint and Fix JavaScript file with StandardJS}
\description{
This addin lints and fixes selected JavaScript code or the currently open
file in RStudio. The addin can be helpful for linting JavaScript code
embedded in R Markdown or Shiny apps, in addition to linting whole JavaScript
files. The underlying functions are not exported from \pkg{js4shiny}. If you
want to programmatically lint multiple files, it would be better to use \code{npm}
scripts or another JavaScript task running system to lint your files.
}
\section{Installing StandardJS}{
 \href{https://standardjs.com/}{standardjs} is a
style guide, code linter, and beautifier in one. It is also a command line
tool (\code{standard}) for automatically formatting JavaScript code in the
\href{https://standardjs.com/}{JavaScript Standard Style}. The command line tool
will also alert users to common style and programmer errors.

Using \code{standard} and this addin requires that \code{node}, \code{npm}, and \code{standard}
be installed on your system. To install \code{node} and \code{npm}, you need to
install Node.js (they come together). Follow \href{https://nodejs.org/en/download/}{the instructions from Node.js} to install these tools. Confirm
that your installation was successful by running \code{npm -v} in a new terminal
session. Once \code{npm} is available, install \code{standard} globally by running
this command in the terminal.\preformatted{npm install standard --global
}
}

\references{
https://standardjs.com/
}
