% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdn.R
\name{mdn_search}
\alias{mdn_search}
\alias{mdn_gadget}
\title{Search the Mozilla Developers Network}
\usage{
mdn_search(term, browse = TRUE, topics = c("js", "api", "css", "html",
  "svg"), locale = "en-US")

mdn_gadget(browse = TRUE, locale = NULL)
}
\arguments{
\item{term}{Search term}

\item{browse}{Should the search results be opened in a browser window? If
not, the URL is returned instead.}

\item{topics}{A selection of topics to search, choosing from `"js"`, `"api"`,
`"css"`, `"html"`, `"svg"`, `"mobile"`, `"canvas"`, `"webdev"`, and
`"standards"`.}

\item{locale}{The locale string for the search query. Default is `"en-US"`,
or set to `NULL` to search MDN without specifying a locale.}
}
\value{
The search URL
}
\description{
Searches the MDN Web Docs for a search term. Includes an RStudio addin for
quick and seamless searching: highlight a term in the source code and run the
addin to search MDN. If no text is highlighted, the addin opens a Shiny
gadget for you to enter your search term.
}
\section{Functions}{
\itemize{
\item \code{mdn_gadget}: A Shiny gadget for searching the MDN docs.
}}

