% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{register_knitr}
\alias{register_knitr}
\alias{register_knitr_output_hooks}
\alias{register_knitr_js_engine}
\title{Register js4shiny knitr components}
\usage{
register_knitr_output_hooks()

register_knitr_js_engine()
}
\description{
Register the js4shiny knitr JavaScript engine or the output hooks. Generally,
you will not need to use these. Instead, see [html_document_js()] or
[html_setup()] for methods that cover most use-cases.
}
