% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{snippets_install}
\alias{snippets_install}
\alias{snippets}
\title{Install js4shiny snippets}
\usage{
snippets_install(install_path = NULL, update = TRUE)
}
\arguments{
\item{install_path}{Where should the snippets be installed? If \code{NULL}, the
snippets will install to a default path based on the current version of
RStudio.}

\item{update}{Should existing snippets be updated in place if there are any
conflicts? Default is yes (\code{TRUE}). Otherwise, new snippets are appended to
the end of the existing file, ensuring that you can recover your previous
snippets by editing the snippets file.}
}
\description{
This function installs a set of R, HTML, JavaScript and CSS snippets that
are helpful when developing Shiny apps and doing web development work in
RStudio. By default, the snippets are installed where RStudio will find
them. If you haven't previously installed snippets to RStudio, these
snippets will mask some of the built-in snippets that ship with RStudio.
}
\section{Updating Existing Snippets}{
 If you already have snippets installed,
you can  you can have the installed snippets update the existing snippets
in place with \code{update = TRUE}. Or you can append the new snippets to the
existing snippets files with \code{update = FALSE}. This option is desirable if
you want to make sure that no snippets are overwritten. The newer snippets
will mask older snippets, but no data will be lost.
}

\examples{
snip_tmp <- tempfile("snippets")
dir.create(snip_tmp)
snippets_install(snip_tmp)

}
