% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{register_knitr}
\alias{register_knitr}
\alias{register_knitr_output_hooks}
\alias{register_knitr_js_engine}
\title{Register js4shiny knitr components}
\usage{
register_knitr_output_hooks(set = TRUE, chunk_hook = NULL)

register_knitr_js_engine(set = TRUE)
}
\arguments{
\item{set}{If \code{FALSE} the output hook or JS engine are returned rather than
setting via knitr directly.}

\item{chunk_hook}{Chunk hook to be applied \emph{after} the js4shiny chunk hook
is applied to the chunk output. If \code{NULL}, then the current chunk hook
is used. Only applies when \code{set = TRUE}.}
}
\description{
Register the js4shiny knitr JavaScript engine or the output hooks. Generally,
you will not need to use these. Instead, see \code{\link[=html_document_js]{html_document_js()}} or
\code{\link[=html_setup]{html_setup()}} for methods that cover most use-cases.
}
